ajustes

+++ Students/AddStudentRequest.cs
@@ -1,4 +1,4 @@
 ﻿namespace ApiCrud.Students
 {
-    public record AddStudentRequest(string Name, sbyte Age, string Email, string Phone, string Address, string Cpf);
+    public record AddStudentRequest(string Name, int Age, string Email, string Phone, string Address, string Cpf);
 }
